% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eea39_countries.R
\docType{data}
\name{eea39_countries}
\alias{eea39_countries}
\title{Country names of the EEA39 countries.}
\format{
A character vector of length 39.
}
\source{
\url{https://www.eea.europa.eu/countries-and-regions}
}
\usage{
eea39_countries
}
\description{
A character vector of length 39 containing the names of the EEA39 countries.
}
\keyword{datasets}
