% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eumohp_plot.R
\name{eumohp_plot}
\alias{eumohp_plot}
\title{Plotting the clipped EU-MOHP data}
\usage{
eumohp_plot(.eumohp_starsproxy, ...)
}
\arguments{
\item{.eumohp_starsproxy}{A list of stars proxy objects as derived
from the function eumohp_clip().}

\item{...}{Additional arguments.}
}
\value{
...
}
\description{
Plots the clipped EU-MOHP data as grid using ggplot2.
}
\examples{
\dontrun{
eumohp_clip(
   directory_input = "directory/to/EU-MOHPfiles/",
   region_name_spatcov = c("italy2"),
   hydrologic_order = 1:4,
   abbreviation_measure = c("dsd", "lp"),
   eumohp_version = "v013.1.1"
) |>
eumohp_plot()


# If you want to plot faster, you can increase the argument downsample:

eumohp_clip(
   directory_input = "directory/to/EU-MOHPfiles/",
   region_name_spatcov = c("italy2"),
   hydrologic_order = 1:4,
   abbreviation_measure = c("dsd", "lp"),
   eumohp_version = "v013.1.1"
) |>
eumohp_plot(downsample = 200)
}
}
