% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eumohp_plot.R
\name{eumohp_plot}
\alias{eumohp_plot}
\title{Plot the clipped EUMOHP data as grid}
\usage{
eumohp_plot(.eumohp_starsproxy, ...)
}
\arguments{
\item{.eumohp_starsproxy}{A list of stars proxy objects as derived
from the function eumohp_clip().}

\item{...}{Additional arguments.}
}
\value{
...
}
\description{
Plot the clipped EUMOHP data as grid
}
\examples{
1 + 1
}
