% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eumohp_write.R
\name{eumohp_write}
\alias{eumohp_write}
\title{Write the clipped EUMOHP data as raster data to GEOTIFF files}
\usage{
eumohp_write(eumohp_starsproxy, directory_output, parallel = FALSE)
}
\arguments{
\item{eumohp_starsproxy}{A number.}

\item{directory_output}{A number.}

\item{parallel}{A number.}
}
\value{
...
}
\description{
Write the clipped EUMOHP data as raster data to GEOTIFF files
}
\examples{
1 + 1
}
